backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
asyncpg==0.29.0

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Security and encryption
cryptography==41.0.7
bcrypt==4.1.2
pyjwt==2.8.0

# Data processing and analysis
pandas==2.1.3
sqlglot==27.0.0
numpy==1.24.3

# AI and OpenAI
openai==1.3.7
aiohttp==3.9.1

# WebSocket support
websockets==12.0

# Caching
diskcache==5.6.3

# HTTP client
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0 